requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.3
Django==1.11.1
idna==2.5
lxml==3.8.0
numpy==1.13.0
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
pytz==2017.2
requests==2.17.3
six==1.10.0
urllib3==1.21.1
